requirements.txt
asgiref==3.4.1
backports.entry-points-selectable==1.1.0
certifi==2021.5.30
cffi==1.14.6
chardet==3.0.4
charset-normalizer==2.0.4
cryptography==3.4.8
defusedxml==0.7.1
distlib==0.3.2
Django==3.2.7
django-allauth==0.45.0
django-crispy-forms==1.12.0
django-jalali==4.3.0
easy-thumbnails==2.7.1
filelock==3.0.12
googletrans==3.1.0a0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2021.9.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
jdatetime==3.6.2
oauthlib==3.1.1
Pillow==8.3.2
pipenv==2021.5.29
platformdirs==2.3.0
psycopg2-binary==2.8.3
pycparser==2.20
PyJWT==2.1.0
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
sqlparse==0.4.1
django-ckeditor==6.1.0
urllib3==1.26.6
virtualenv==20.7.2
virtualenv-clone==0.5.6
